Added .nsvf generation as synteshis post-step
Since we are currently using .nsvf files for deployments, added its
generation as synteshis post-step.

NAT's svf_to_nsvf-linux-x86.bin is required and has to be on PATH. If
it's not found, only the .svf file is generated.

NOTE: It is assumed that FPGA is the only tap on its JTAG daisy chain.