FOFB Processing co-simulation
Now it is possible to integrate the fofb_processing controller into
higher level simulations, so it is possible to simulate in a closed
loop configuration using a machine model.

You can build and run the fofb_processing_cosim server using hdlmake
and make. You should have installed an up-to-date Rust toolchain in
your system and GHDL built with llvm or gcc backends (mcode will not
work):

$ cd hdl/testbench/fofb_processing_cosim/ghdl
$ hdlmake
$ make
$ ./fofb_processing_cosim -gg_TCP_PORT=14000