fofb_processing/cheby: clean registers api
By the time this layout was drawn, cheby had a bug that didn't allow the
usage of 'repeat' statements if the design contained memories. The
workaround was to flatenize the per-channel registers/memories
declarations.

The issue mentioned above doesn't appear on cheby 1.4.0. So, this commit
cleans up the registers api.