Change BPM position error calculation to 'sp - pos'
During tests with the new FOFB IOC we observed that all correctors
would saturate immediately upon closing the loop, suggesting a
positive feed-back somewhere, and after some debugging we found that
it was occurring due way the BPM position error was computed by the
fofb_processing core.

Conventionally, the error in a control system is calculated as
'reference - measurement', and we assumed this in the upper software
stack. But the fofb_processing core was computing the error as
'measurement - reference' instead, and this behavior was masked by a
software bug that effectively multiplied all inverse matrix
coefficients by '-1'.

This is a breaking change, so I incremented the xwb_fofb_processing
ABI major number version to reflect it.