Use a local version of sim_wishbone.vhd instead of the general-cores one
The original sim_wishbone.vhd implementation puts the wishbone address
line in the 'Unknown' state ('X') after receiving an ack from the
slave. I don't know why this is done, but it causes problems with RAM
instances on cheby generated VHDL (lots of 'metavalue detected'
warnings).

I commented this line and set the wb_o.adr signal to zeros in the
init() procedure.