fofb_shaper_filt: use number of biquads to define order
Instead of defining the maximum order, use a generic to define the
number of internal biquads (the order is twice this value). This
simplifies things since we can use it directly (i.e. no intermediate
computation) for defining internal dimensions, number of iterations etc.

NOTE: The ABI was broken. The register 'max_filt_order' was changed to
      'num_biquads'.